requirements.txt
Django==2.1.5
pytz==2018.9
graphene-django>=2.0
psycopg2-binary
django-autofixture
coverage
